break services out of main
