+++ rocksdb-raft/src/store/mod.rs
@@ -1,4 +1,4 @@
-pub mod store;
+pub mod storage;
 pub mod log_storage;
 pub mod state_machine;
 pub mod types;
